Separate pb_file and validation. Add CLI entry point for GUI.
